requirements.txt
Flask==3.1.3
gunicorn==25.3.0
keras==3.14.0
lime==0.2.0.1
matplotlib==3.10.9
numpy==2.4.4
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
pandas==3.0.2
Pillow==12.2.0
scikit_learn==1.8.0
seaborn==0.13.2
shap==0.51.0
skimage==0.0
tensorflow==2.21.0
